fix type errors + add changesets
