+++ .changeset/chatty-rocks-pull.md
@@ -1,0 +1,5 @@
+---
+'@ai-sdk/alibaba': patch
+---
+
+feat(provider/alibaba): initial alibaba provider
