+++ .changeset/twenty-masks-grow.md
@@ -1,0 +1,5 @@
+---
+'@ai-sdk/openai-compatible': patch
+---
+
+feat(provider/alibaba): initial alibaba provider
